pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "agenttalk"
version = "1.0.0"
description = "Production-grade real-time messaging platform for humans and AI agents"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlalchemy==2.0.35",
    "greenlet==3.1.1",
    "asyncpg==0.29.0",
    "alembic==1.13.3",
    "redis==5.1.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "pydantic[email]==2.9.2",
    "email-validator==2.2.0",
    "pydantic-settings==2.5.2",
    "python-multipart==0.0.12",
    "httpx==0.27.2",
]

[project.optional-dependencies]
test = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "httpx==0.27.2",
    "aiosqlite==0.20.0",
]
dev = [
    "ruff==0.6.9",
    "mypy==1.11.2",
    "pre-commit==3.8.0",
]

# ── pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--tb=short",
    "--strict-markers",
    "-q",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring live services",
]

# ── coverage ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["app"]
omit = [
    "app/db/session.py",   # engine config, not worth unit testing
    "alembic/*",
]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 70

# ── ruff (linter + formatter) ─────────────────────────────────────────────────
[tool.ruff]
target-version = "py312"
line-length = 100
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = [
    "E501",   # line too long — handled by formatter
    "B008",   # do not perform function calls in default arguments (FastAPI Depends)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ── mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
warn_return_any = false
